Pass the shell arguments directly to bash
This allows us to run the "run" script using "set -eo pipefail" by default.

The previous wrapper executed the arguments as a bash command string, which
allowed the user to run any command, but didn't give us a chance to special case
if the argument was a script passed from the GHA "run" command.

This is a breaking change.

Porting guide:

* Instead of `msys2 mybinary` you now have to run `msys2 -c 'mybinary'`
* To revert the bash error settings you can use `set +e` and `set +o pipefail`
  in your run script