Rework pacman pkg cache and enable by default
Three main improvements:

* Create a separate cache for each input configuration: Assuming a build matrix
  which builds for different arches and has different dependencies we create
  a cache for each of them, while loading any of them.
* Prune caches before saving them: Every time a package has a new version we
  create a new cache which only includes the new packages. This makes sure that
  the cache/restore size stays as small as possible over time.
* Avoid cache race conditions: If a run doesn't change the cache we wont save it
  and if saving fails because another job created it in the mean time,
  we catch the error and ignore it.

Overall this cache doesn't save much time, since installation and initial setup
take the most time, but this should save a lot of traffic for our main
repo server.

And also this removes the cache option which was confusing some users because it
only caches packages and not everything.