Cache the whole installation when possible
In case release=true we try to cache the whole installation.

As cache restore key we use the whole input + the installer checksum, so
we only get a similar setup back.

For saving we hash the content of /var/lib/pacman/local which contains all the info
about which packages are installed.

Fixes #50