Handle pacman upgrade creating .pacnew files
In case a pacman upgrade changes a file that is marked for backup
it leaves it alone and installs it as <name>.pacnew instead and warns
the user.

In our case we change pacman.conf to skip CheckSpace. If pacman would
then update itself it would save any new pacman.conf as pacman.conf.pacnew.
Fix this by replacing the file manually.