Don't abort when run on Linux / Mac OS.
I am trying to call this action from another composite action, one that
is supposed to be working for Linux, Mac OS, and Windows.

Ideally one would put the `uses` statement behind a conditional
expression, like:

    - uses: msys2/setup-msys2@v2
      if: runner.os == 'Windows'

However, composite actions do not support conditional steps yet (see:
https://github.com/actions/runner/issues/834), so the only way to keep
the setup-msys2 step inside my action (and not forcing the wokflow to
use it instead), is to suppress the error message (It's OK to log it,
but it should not fail).