Add an output parameter for the install location
This adds the "msys2-location" output parameter which contains the
absolute Windows path of the MSYS2 installation that was used by
the action.

While cygpath can be used as well, some users might just want to use
the path to pass it to some other tool and don't want to run any msys2
things themselves.